+++ py_md_doc/__init__.py
@@ -1,2 +1,3 @@
 from .py_md_doc import PyMdDoc
+from .class_inheritance import ClassInheritance
 from .parameter import Parameter